separated configuration reader and validator modules
