+++ src/configuration-reader.js
@@ -1,5 +1,6 @@
 const fs = require('fs');
 const path = require('path');
+const Validator = require('./configuration-validator');
 
 module.exports = function ConfigurationReader() {
   const defaultConfiguration = {
@@ -27,30 +28,8 @@ module.exports = function ConfigurationReader() {
     }
 
     const configuration = require(absoluteConfigPath);
-    validate(configuration);
+    (new Validator()).validate(configuration);
 
     return configuration;
   };
-
-  this.validate = validate;
-
-  function validate(configuration) {
-    configuration.environments.forEach((environment) => {
-      if (typeof environment.extension === 'undefined' || environment.extension === '') {
-        throw new TypeError('extension must be defined for each environment');
-      }
-
-      if (typeof environment.testNameSuffix === 'undefined' || environment.testNameSuffix === '') {
-        throw new TypeError('testNameSuffix must be defined for each environment');
-      }
-
-      if (typeof environment.testDir === 'undefined' || environment.testDir === '') {
-        throw new TypeError('testNameSuffix must be defined for each environment');
-      }
-
-      if (typeof environment.testRunnerCommand === 'undefined' || environment.testRunnerCommand === '') {
-        throw new TypeError('testRunnerCommand must be defined for each environment');
-      }
-    });
-  }
 };
