+++ src/configuration-validator.js
@@ -1,0 +1,21 @@
+module.exports = function ConfigurationValidator() {
+  this.validate = (configuration) => {
+    configuration.environments.forEach((environment) => {
+      if (typeof environment.extension === 'undefined' || environment.extension === '') {
+        throw new TypeError('extension must be defined for each environment');
+      }
+
+      if (typeof environment.testNameSuffix === 'undefined' || environment.testNameSuffix === '') {
+        throw new TypeError('testNameSuffix must be defined for each environment');
+      }
+
+      if (typeof environment.testDir === 'undefined' || environment.testDir === '') {
+        throw new TypeError('testNameSuffix must be defined for each environment');
+      }
+
+      if (typeof environment.testRunnerCommand === 'undefined' || environment.testRunnerCommand === '') {
+        throw new TypeError('testRunnerCommand must be defined for each environment');
+      }
+    });
+  };
+};
